backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0

supabase==2.3.0
httpx==0.24.1

asyncpg==0.29.0
psycopg2-binary==2.9.9
pgvector==0.2.4

python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

pydantic==2.5.2
pydantic-settings==2.1.0
python-dotenv==1.0.0

cohere==4.37.0

PyPDF2==3.0.1
python-docx==1.1.0